+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/datanode/erasurecode/ErasureCodingWorker.java
@@ -253,7 +253,7 @@ public final class ErasureCodingWorker {
     private long positionInBlock;
 
     // sources
-    private final short[] liveIndices;
+    private final byte[] liveIndices;
     private final DatanodeInfo[] sources;
 
     private final List<StripedReader> stripedReaders;
